Load WebAPI from root of site
Calls to get JSON data from the WebAPI endpoints were using url's
relative to the page url.  that worked fine for the default route, but
with the new Home/Inline view, the relative path becomes
Home/Inline/api/temperature, which can't be found.  This changes the
function which makes the REST call to load webapi from the root of the
site.  This will work unless the site is deployed to a subdirectory.